♻️Refactor: 삭제 시 범위를 벗어나면 에러를 던지도록 수정

+++ PDFViewer/PDFViewer/Presentation/PDFList/ViewModel/PDFListViewModel.swift
@@ -62,6 +62,10 @@ extension DefaultPDFListViewModel {
   }
   
   func deletePDF(at index: Int) throws {
+    guard (0..<pdfDatas.count).contains(index) else {
+      throw RepositoryError.deletionFailed
+    }
+    
     let deletePDFData = pdfDatas[index]
     
     try repository.deletePDFData(pdfData: deletePDFData)
